Add connection ID #cN for client lifecycle traceability
Every client connection gets a monotonic ID assigned at TCP accept time.
Format: #cN (e.g., #c42). Appears in all client-side log messages:

  [app@mydb #c42] client connected from 10.0.0.5:41234 (TLS)
  [app@mydb #c42] acquired server pid=98765
  [app@mydb #c42] DEALLOCATE ALL: cleared 5 entries
  [app@mydb #c42] disconnected, session=5m12s

DBA can now trace a client's full lifecycle: grep '#c42\]' pg_doorman.log

Replaces random client_id (i32) with monotonic connection_id (u64).
Cancel Protocol process_id derived as connection_id as i32 (sequential
gives fewer collisions than random at <50K concurrent clients).

SHOW CLIENTS displays #cN format instead of hex process_id.

Changes: 10 files. connection_id field added to Client struct and
ClientStats. Threaded through client_entrypoint → startup → Client.
64 log format strings updated to include #c{connection_id}.

+++ src/stats/mod.rs
@@ -56,7 +56,7 @@ pub use socket::get_socket_states_count;
 // -----------------------------------------------------------------------------
 /// Type alias for the client statistics lookup table.
 /// Maps client IDs to their corresponding statistics objects.
-type ClientStatesLookup = HashMap<i32, Arc<ClientStats>>;
+type ClientStatesLookup = HashMap<u64, Arc<ClientStats>>;
 
 /// Type alias for the server statistics lookup table.
 /// Maps server IDs to their corresponding statistics objects.
@@ -119,7 +119,7 @@ impl Reporter {
     ///
     /// If a client with the same ID is already registered, a warning is logged and
     /// the registration is ignored to prevent overwriting existing statistics.
-    fn client_register(&self, client_id: i32, stats: Arc<ClientStats>) {
+    fn client_register(&self, client_id: u64, stats: Arc<ClientStats>) {
         use std::collections::hash_map::Entry;
         match CLIENT_STATS.write().entry(client_id) {
             Entry::Occupied(_) => {
@@ -139,7 +139,7 @@ impl Reporter {
     /// # Arguments
     ///
     /// * `client_id` - Unique identifier for the client to unregister
-    fn client_disconnecting(&self, client_id: i32) {
+    fn client_disconnecting(&self, client_id: u64) {
         CLIENT_STATS.write().remove(&client_id);
     }
